[#8] Refactor using the component framework
The project is now using the component framework. This
implies that the whole system can be started and stopped
in a simple way. Each component have their own public API
for other components to use. This commit also removes the
tight connection with using the lein-ring plugin in favour
of go, stop, reset functions in dev/user.clj.